[IMP] product_medical: views for conformity categories
The attributes for the Conformity Configuration (Medical Class,
Medical Category, PPE Category, and In Vitro Diagnostics) had no
views associated to them, so the only way of accessing them was
through the entry they have in the form for product.template.

With this PR, tree views and their actions have been added, so
that they can be reused in other modules to accessed them
from custom menu-items.